+++ docs/components/ExamplePage2.vue
@@ -6,7 +6,7 @@ import IsInsideIndicator from './IsInsideIndicator.vue';
 
 <template>
     <div class="p-4 min-h-screen flex">
-        <div class="border-4 border-dashed p-4 w-full border-gray-500">
+        <div class="border-4 border-dotted p-4 w-full border-orange-500">
             <h1 class="text-3xl font-semibold">Interactions</h1>
             <IsInsideIndicator />
 
@@ -15,16 +15,7 @@ import IsInsideIndicator from './IsInsideIndicator.vue';
                     Close
                 </button>
                 <button
-                    @click="
-                        () =>
-                            updateStyles({
-                                width: '90%',
-                                height: '90%',
-                                radius: 20,
-                                backdropOpacity: 0.7,
-                                backdropColor: '#f26f25',
-                            })
-                    "
+                    @click="() => updateStyles({ width: '90%', height: '90%', radius: 20, backdropOpacity: 0.7 })"
                     class="bg-[#3c3c43] text-[#fffff5db] px-4 py-1 rounded-md"
                 >
                     Update Styles
@@ -33,10 +24,22 @@ import IsInsideIndicator from './IsInsideIndicator.vue';
                     Hide Close Button
                 </button>
                 <button
-                    @click="() => createLightbox('http://localhost:5173/lightbox-sdk/example-page-2').open()"
+                    @click="() => createLightbox('http://localhost:5173/lightbox-sdk/example-page').open()"
                     class="bg-[#3c3c43] text-[#fffff5db] px-4 py-1 rounded-md"
                 >
                     Open Other Page
+                </button>
+                <button
+                    @click="() => createLightbox('https://rollupjs.org/tutorial/#installing-rollup-locally').open()"
+                    class="bg-[#3c3c43] text-[#fffff5db] px-4 py-1 rounded-md"
+                >
+                    Open test
+                </button>
+                <button
+                    @click="() => createLightbox('https://rollupjs.org/tutorial/#installing-rollup-locally').close()"
+                    class="bg-[#3c3c43] text-[#fffff5db] px-4 py-1 rounded-md"
+                >
+                    Open test CLOSE
                 </button>
             </div>
         </div>