+++ lib/src/config.ts
@@ -1,5 +1,0 @@
-import { Dimensions } from './constants';
-
-export const Config = {
-    defaultDimension: Dimensions.MD,
-};