+++ lib/src/core/index.ts
@@ -1,43 +1,40 @@
-import { LightboxEvents } from '../constants';
+import { LightboxEvents as LE } from '../constants';
 import { postMessage, redirectBasedOnDriver } from '../helpers';
-import { InitOptions, LightboxInstance, ApiStructure, LightboxStyles, ClientCallback } from '../types';
-import { mountLightbox, mountListener } from './assemblers';
+import { InitialOptions, LightboxInstance, ApiStructure, LightboxStyles, ClientCallback } from '../types';
+import { mountLightbox, unmountLightbox } from './assemblers';
 
 export const isInside = () => globalThis.location !== globalThis.parent.location;
+export const updateStyles = (styles: LightboxStyles) => postMessage(LE.UPDATE_STYLES, styles);
+export const hideCloseButton = () => postMessage(LE.HIDE_CLOSE_BUTTON);
 
-export const emitClose = (payload?: unknown) => postMessage(LightboxEvents.CLOSE, payload);
-
-export const emitUpdateStyles = (styles: LightboxStyles) => postMessage(LightboxEvents.UPDATE_STYLES, styles);
-
-export const emitHideCloseButton = () => postMessage(LightboxEvents.HIDE_CLOSE_BUTTON);
-
-export const emit = ({ type, payload, preventClose }: ApiStructure) => {
-    postMessage(type, payload, preventClose);
-    return { close: emitClose };
+export const emitClose = (payload?: unknown, preventClose?: boolean) => {
+    postMessage(LE.CLOSE, payload, preventClose);
 };
 
-export const createLightbox = (
-    url: string,
-    options?: InitOptions & {
-        launch?: boolean;
-    }
-): LightboxInstance => {
+export const emit = ({ type, payload, preventClose }: ApiStructure) => postMessage(type, payload, preventClose);
+
+export const createLightbox = (url: string, options?: InitialOptions): LightboxInstance => {
     const lightbox: LightboxInstance = {
         allowRedirects: options?.allowRedirects ?? true,
         callbacks: options?.callbacks ?? {},
         closeButton: options?.closeButton ?? true,
         styles: options?.styles ?? {},
         url: url,
-        close: emitClose,
-        updateStyles: emitUpdateStyles,
-        hideCloseButton: emitHideCloseButton,
+        close: () => unmountLightbox(url),
+        updateStyles,
+        hideCloseButton,
         on: (name: string, callback: ClientCallback) => {
             lightbox.callbacks[name] = callback;
         },
         open: () => {
             if (lightbox.allowRedirects) redirectBasedOnDriver(url);
-            mountListener(lightbox.callbacks, lightbox.styles);
-            mountLightbox(url, lightbox.styles, lightbox.closeButton);
+            mountLightbox({
+                url,
+                callbacks: lightbox.callbacks,
+                styles: lightbox.styles,
+                closeButtonEnabled: lightbox.closeButton,
+                enforceStyles: options?.enforceStyles ?? false,
+            });
         },
     };